API/requirements.txt
apispec==5.2.2
apispec_webframeworks==0.5.2
Flask==2.1.0
flask_swagger_ui==4.11.1
marshmallow==3.17.0
matplotlib==3.4.3
numpy==1.20.3
Pillow==9.1.1
tensorflow==2.7.0
